Move cmd/mcp_cmds.go to pkgs/prompts/cost/cost.go
